controlFuzzy implementado

+++ firmware/RobotSerialCommunication/agent/common.h
@@ -13,7 +13,7 @@ const int HEADER_LEN (sizeof(unsigned short)*3);
 
 
 struct appdata {
-  int InitFlag;
+  unsigned short InitFlag;
   unsigned short id;
   unsigned short op; //codigo de operacion
   unsigned short len;                       /* longitud de datos */
